requirements.txt
streamlit~=1.52.2
transformers~=4.57.3
pandas~=2.3.3
pillow~=12.1.0
plotly